backend/requirements.txt
absl-py
aiohttp
aiosignal
annotated-types
anyio
api-display-purposes
asgiref
astunparse
attrs
awscli
backoff
bcrypt
beautifulsoup4
blis
boto3
botocore
build
cachetools
catalogue
certifi
chardet
charset-normalizer
chroma-hnswlib
chromadb
clarifai
clarifai-grpc
click
cloudpathlib
colorama
coloredlogs
confection
configparser
contextlib2
contourpy
cycler
cymem
dataclasses-json
Deprecated
distro
dnspython
docutils
EasyProcess
email_validator
emoji
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi
fastapi-cli
filelock
flatbuffers
fonttools
frozenlist
fsspec
future
gast
gitdb
GitPython
google-api-core
google-auth
google-cloud-language
google-pasta
googleapis-common-protos
grpcio
grpcio-status
h11
h5py
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
iniconfig
instapy
Jinja2
jmespath
jsonpatch
jsonpointer
jsonschema
keras
kiwisolver
kubernetes
langchain
langchain-cli
langchain-community
langchain-core
langchain-text-splitters
langcodes
langserve
langsmith
language_data
lark
libclang
libcst
marisa-trie
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
MeaningCloud-python
ml-dtypes
mmh3
monotonic
mpmath
multidict
murmurhash
mypy-extensions
namex
networkx
numpy
oauthlib
onnxruntime
openai
opencv-python
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
opt-einsum
optree
orjson
outcome
overrides
packaging
pandas
pillow
pluggy
plyer
posthog
preshed
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydot
Pygments
pyparsing
pypdf
PyPika
pyproject-toml
pyproject_hooks
PySocks
pytest
python-dateutil
python-dotenv
python-multipart
python-rapidjson
python-telegram-bot
pytz
PyVirtualDisplay
PyYAML
regex
requests
requests-oauthlib
rich
rsa
s3transfer
safetensors
schedule
schema
selenium
semantic-version
setuptools-rust
shellingham
six
smart-open
smmap
sniffio
sortedcontainers
soupsieve
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
sse-starlette
starlette
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
tf_keras
thinc
tokenizers
toml
tomlkit
torch
tqdm
transformers
trio
trio-websocket
tritonclient
typer
typing-inspect
typing_extensions
tzdata
ujson
urllib3
uvicorn
uvloop
wasabi
watchfiles
weasel
webdriver-manager
webdriverdownloader
websocket-client
websockets
Werkzeug
wrapt
wsproto
yarl
zipp
